requirements.txt
# RetinaDA项目依赖
# 基础科学计算和数据处理
numpy>=1.19.0
pandas>=1.1.0
scipy>=1.5.0
scikit-learn>=0.23.0

# 深度学习框架
torch>=1.8.0
torchvision>=0.9.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0

# 可视化和评估
tqdm>=4.50.0